\name{BreastCancerRelatedGenes}
\alias{BreastCancerRelatedGenes}
\title{Breast cancer related genes affected by mutations curated from published papers.}
\description{
	Breast cancer related genes affected by mutations curated from published papers.
}
\usage{data(BreastCancerRelatedGenes)}
\format{
A vector of breast cancer-related genes affected by mutations, which are collected from published papers.
}
\source{
1. Ciriello, G., Gatza, M. L., Beck, A. H., Wilkerson, M. D., Rhie, S. K., Pastore, A., … Perou, C. M. (2015). Comprehensive Molecular Portraits of Invasive Lobular Breast Cancer. Cell, 163(2), 506–519. doi:10.1016/j.cell.2015.09.033 Figure 1B and Table 1

2. Cancer, T., & Atlas, G. (2012). Comprehensive molecular portraits of human breast tumours. Nature, 490(7418), 61–70. doi:10.1038/nature11412 Figure 1

3. Ian R. Watson et al. Emerging patterns of somatic mutations in cancer. Nature Reviews Genetics 14, 703–718 (2013) doi:10.1038/nrg3539
}
\references{}
\keyword{datasets}
