\name{CancerDB}
\alias{CancerDB}
\title{Cancer related genes and pathways curated from published papers.}
\description{
	Cancer related genes and pathways curated from published papers.
}
\usage{data(CancerDB)}
\format{
A list of variables:\cr
CCSP:   Cancer cell signalling pathways\cr
Smg127: SMGs identified in and across 12 human cancer types.\cr
CDG435: Cancer driver genes\cr
CGC:    Cancer Gene Census\cr
}
\source{
1. Cyriac Kandoth et. al  Mutational landscape and significance across 12 major cancer types  doi:10.1038/nature12634

Figure2 | The 127 SMGs from 20 cellular processes in cancer identified in 12 cancer types

2. Bert Vogelstein et al. Cancer Genome Landscapes  Science 339, 1546 (2013);DOI: 10.1126/science.1235122

Fig. 7. Cancer cell signaling pathways and the cellular processes they regulate. Curated from Table S2A.

3. David Tamborero et. al  Comprehensive identification of mutational cancer driver genes across 12 tumor types. DOI: 10.1038/srep02650

Supplementary Information (2.58 KB), Supplementary Table 2.

}
\references{}
\keyword{datasets}
