\name{CancerGeneCensus}
\alias{CancerGeneCensus}
\title{Cancer Gene Census.}
\description{
	Cancer Gene Census data (Download date: 2014-03-19).
}
\usage{data(CancerGeneCensus)}
\format{
A list of 3 elements, including 1 data.frames of Cancer Gene Census data, link and download date.
}
\source{
http://cancer.sanger.ac.uk/cancergenome/projects/census/
}
\references{}
\keyword{datasets}
