\name{CancerRelatedGenes}
\alias{CancerRelatedGenes}
\title{Cancer related genes affected by mutations curated from published papers.}
\description{
	Cancer related genes affected by mutations curated from published papers.
}
\usage{data(CancerRelatedGenes)}
\format{
A vector of cancer-related genes affected by mutations, which are collected from published papers.
}
\source{
1. Cancer Gene Census:
http://cancer.sanger.ac.uk/cancergenome/projects/census/

2. Cyriac Kandoth et. al  Mutational landscape and significance across 12 major cancer types  doi:10.1038/nature12634
Supplementary Table 4.  Significantly mutated genes in all cancer types and 12 separate cancer types

3. David Tamborero et. al  Comprehensive identification of mutational cancer driver genes across 12 tumor types. DOI: 10.1038/srep02650
Supplementary Information, Supplementary Table 2

4. Bert Vogelstein et al. Cancer Genome Landscapes  Science 339, 1546 (2013);DOI: 10.1126/science.1235122
Fig. 7. Cancer cell signaling pathways and the cellular processes they regulate. Curated from Table S2A.
}
\references{}
\keyword{datasets}
