\name{ChromatinRegulatoryGenes}
\alias{ChromatinRegulatoryGenes}
\title{Chromatin regulatory genes implicated in human cancers curated from published papers.}
\description{
	Chromatin regulatory genes implicated in human cancers curated from published papers.
}
\usage{data(ChromatinRegulatoryGenes)}
\format{
A list of multiple variables.
}
\source{

1. Abel Gonzalez-Perez, Alba Jene-Sanz and Nuria Lopez-Bigas  The mutational landscape of chromatin regulatory factors across 4,623 tumor samples  Genome Biology 2013, 14:r106

Figure 1 Likely driver chromatin regulatory factors across the datasets of somatic mutations in IntOGen-mutations.

Figure 3 FM bias, mutation frequencies and mutually exclusivity of chromatin regulatory factors as part of complexes.

}
\references{}
\keyword{datasets}
