\name{get.mutland}
\alias{get.mutland}
\title{Prepare data for \code{plot.landscape}}
\description{
Prepare mutational landscape data for \code{plot.landscape}. Automatically selected q-value cutoff to nominate SMGs.
}
\usage{
get.mutland(genes=NULL, mutsigcv.sig.gene.file, gene.mut.freq.file, mut.per.smp.file, gene.mut.mat.file, use.q.value=TRUE, discard.genes=c("TTN"))
}
\arguments{
  \item{genes}{A vector of gene symbols to be selected, if NULL automatically select SMGs.}
  \item{mutsigcv.sig.gene.file}{MutSigCV output file, often suffixed with \code{*.sig_genes.txt}.}
  \item{gene.mut.freq.file}{Gene mutation frequency file.}
  \item{mut.per.smp.file}{Mutation per sample file.}
  \item{gene.mut.mat.file}{Gene mutation matrix file.}
  \item{use.q.value}{Use q-value in \code{mutsigcv.sig.gene.file}, else use p-value.}
  \item{discard.genes}{Gene to be ignored in \code{genes}.}
}
\details{
\code{mutsigcv.sig.gene.file} is generated by \code{MutSigCV}.

\code{gene.mut.freq.file}, \code{mut.per.smp.file}, \code{gene.mut.mat.file} can be generaed by \code{ngstk1.0 muMat}.
}
\seealso{
\code{plot.landscape}, \code{plot.landscape2}
}
\examples{
# Not run
r=get.mutland(genes=NULL, "../MutSigCV/GC.WES.sig_genes.txt", "../basic_stats/WES.exonic.gene_mut_freq.txt", "../basic_stats/WES.exonic.mut_per_sample.txt", "../basic_stats/WES.exonic.gene_mut_mat.txt", F, top.k.gene=13, discard.genes="TTN")
}

\keyword{utilities}
