\name{HDIofMCMC_for_stanfit}
\alias{HDIofMCMC_for_stanfit}
\title{HDI for stanFit object}
\description{
Calculate high density interval (HDI) for an stanFit object.
}
\usage{
HDIofMCMC_for_stanfit(stanFit, credMass=0.95)
}
\arguments{
  \item{stanFit}{an object of stanFit. E.g. returned by \code{best.robust_t_test(...)}.}
  \item{credMass}{confidence intervals}
}
\details{

}
\seealso{
\code{best.robust_t_test},\code{stan_model},\code{HDIofMCMC}
}
\examples{
}

\keyword{utilities}
