\name{landscape.data}
\alias{landscape.data}
\title{Prepare data for \code{plot.landscape}}
\description{
Prepare mutational landscape data for \code{plot.landscape}.
}
\usage{
landscape.data(genes=NULL, mutsigcv.sig.gene.file, gene.mut.freq.file, mut.per.smp.file, gene.mut.mat.file, use.q.value=T, top.k.gene=12, discard.genes=c("TTN"))
}
\arguments{
  \item{genes}{A vector of gene symbols to be selected}
  \item{mutsigcv.sig.gene.file}{MutSigCV output file, often suffixed with \code{*.sig_genes.txt}.}
  \item{gene.mut.freq.file}{Gene mutation frequency file.}
  \item{mut.per.smp.file}{Mutation per sample file.}
  \item{gene.mut.mat.file}{Gene mutation matrix file.}
  \item{use.q.value}{Use q-value in \code{mutsigcv.sig.gene.file}, else use p-value.}
  \item{top.k.gene}{The top of highly mutated gene count. If \code{genes} is NULL use the \code{top.k.gene} highly mutated genes.}
  \item{discard.genes}{Gene to be ignored in \code{genes}.}
}
\details{
\code{mutsigcv.sig.gene.file} is generated by \code{MutSigCV}.

\code{gene.mut.freq.file}, \code{mut.per.smp.file}, \code{gene.mut.mat.file} can be generaed by \code{ngstk1.0 muMat}.
}
\seealso{
\code{plot.landscape}
}
\examples{
# Not run
r=landscape.data(genes=NULL, "../MutSigCV/GC.WES.sig_genes.txt", "../basic_stats/WES.exonic.gene_mut_freq.txt", "../basic_stats/WES.exonic.mut_per_sample.txt", "../basic_stats/WES.exonic.gene_mut_mat.txt", F, top.k.gene=13, discard.genes="TTN")
}

\keyword{utilities}
