\name{LandscapeColor}
\alias{LandscapeColor}
\title{Multiple color scheme designed for \code{plot.landscape}.}
\description{
	Multiple color scheme designed for \code{plot.landscape} based on RColorBrewer.
}
\usage{data(LandscapeColor)}
\format{null.
}
\source{
http://colorbrewer2.org
}
\references{}
\keyword{datasets}
