\name{legoplot}
\alias{legoplot}
\title{Legoplot exihibition of 96 mutational categories.}
\description{
Legoplot exihibition of 96 mutational categories.
}
\usage{
legoplot(originalGenomes_file, counts=NULL, alpha=0.8, scalez=0.15, gap=0.1, contexts_file=system.file('data/contexts.txt',package='lxctk'), snapshot_fig, piechart_fig)
}
\arguments{
  \item{originalGenomes_file}{Input file generated by decipherMutationalSignatures (https://github.com/lixiangchun/decipherMutationalSignatures).}
  \item{counts}{A numeric vector, i.e. rowSums of \code{originalGenomes_file}. If provided, \code{originalGenomes_file} is ignored.}
  \item{alpha}{Transparency.}
  \item{scalez}{Magnitude scaling of z-axis.}
  \item{gap}{Gap between barplots.}
  \item{contexts_file}{Leave it unchanged.}
  \item{snapshot_fig}{PNG snapshot of legoplot.}
  \item{piechart_fig}{PDF output file for the pie chart.}
}
\details{
See the following example for more info about inputs.
}
\seealso{
\code{...}
}
\examples{
library(lxctk)
legoplot(system.file('data/stad_regular_GC_originalGenomes.txt', package="lxctk"))
}

\keyword{utilities}
