\name{LixcCuratedPathway}
\alias{LixcCuratedPathway}
\title{Cancer related pathways affected by mutations curated from published papers.}
\description{
	Cancer related pathways affected by mutations curated from published papers.
}
\usage{data(LixcCuratedPathway)}
\format{
A data.frame of 3 columns, representing 32 sigalling pathways often affected by mutations.
}
\source{
1. Cyriac Kandoth et. al  Mutational landscape and significance across 12 major cancer types  doi:10.1038/nature12634
Figure2 | The 127 SMGs from 20 cellular processes in cancer identified in 12 cancer types

2. Bert Vogelstein et al. Cancer Genome Landscapes  Science 339, 1546 (2013);DOI: 10.1126/science.1235122
Fig. 7. Cancer cell signaling pathways and the cellular processes they regulate. Curated from Table S2A.
}
\references{}
\keyword{datasets}
