\name{mutsigcv.sig.pathway}
\alias{mutsigcv.sig.pathway}
\title{Identify significantly mutated pathways based on MutSigCV outputs via multiple p-value combination}
\description{
Identify significantly mutated pathways (SMP) through combining p-values that computed for each gene, e.g. output produced by MutSigCV.
}
\usage{
mutsigcv.sig.pathway(mutsigcv.sig_gene.filename, pathway, out.filename=NA, exclude.genes=c(), min.n_nonsilent=0)
}
\arguments{
  \item{mutsigcv.sig_gene.filename}{Output file generated by MutSigCV, often ended with \code{.sig_genes.txt}.}
  \item{pathway}{A data frame with 3 cols containg pathway information, see \code{details} for detail.}
  \item{out.filename}{output file.}
  \item{exclude.genes}{Genes to be excluded.}
  \item{min.n_nonsilent}{min number of nonsilent mutations required.}
}
\details{
Identification of SMPs based on combining p-values.

\code{pathway}, a data frame, with 3 columns, for example:

PathwayID	PathwayName	Genes\cr
pathway1	RB signaling	A,B,C,D,E\cr
pathway2	TP53 signaling	C,D,F,G,H

}
\seealso{
}
\examples{
mutsigcv.sig_gene.filename <- "EC.sig_genes.txt"
pathway.filename <- "/ifshk1/BC_CANCER/03user/lixiangchun/db/pathway/BroadCuratedPathway/classic_cancer_pathway.lxc4"
pathway <- read.table(pathway.filename, sep="\t")
out.filename = "EC.sig_pathways.txt"
mutsigcv.sig.pathway(mutsigcv.sig_gene.filename, pathway, out.filename)
}

\keyword{utilities}
