\name{nature12912_260_Smg}
\alias{nature12912_260_Smg}
\title{260 SMG from nature12912.}
\description{
	260 SMG from nature12912 Supplementary Table 2.
}
\usage{data(nature12912_260_Smg)}
\format{
A vector of 260 SMGs from nature12912.
}
\source{
Michael S. Lawrence et al. Discovery and saturation analysis of cancer genes across 21 tumour types  doi:10.1038/nature12912.
}
\references{}
\keyword{datasets}
