\name{OncogenicSignatureGeneSet}
\alias{OncogenicSignatureGeneSet}
\title{Oncogenic signatures gene sets curated by GSEA.}
\description{
	Oncogenic signatures gene sets curated by GSEA, downloaded from GSEA (v4.0).
}
\usage{data(OncogenicSignatureGeneSet)}
\format{
A data.frame of 3 columns.
}
\source{
http://www.broadinstitute.org/gsea/downloads.jsp
}
\references{}
\keyword{datasets}
