\name{PathwayDB}
\alias{PathwayDB}
\title{Canonical pathways curated by GSEA.}
\description{
	Canonical pathways from KEGG, BIOCARTA and REACTOME, downloaded from GSEA (v4.0).
}
\usage{data(PathwayDB)}
\format{A list of 3 canonical pathways.
}
\source{
http://www.broadinstitute.org/gsea/downloads.jsp
}
\references{}
\keyword{datasets}
