\name{RegularMutatedGC}
\alias{RegularMutatedGC}
\title{SMGs in regular-mutated genes and related clinicopathological information.}
\description{
	SMGs in regular-mutated genes and related clinicopathological information from Li X.C. et al Cancer Research paper.
}
\usage{data(RegularMutatedGC)}
\format{
A data frame of SMGs in regular-mutated genes and related clinicopathological information 
from Li X.C. et al Cancer Research paper.
}
\source{
Li, Xiangchun*, Wu, W. K. K., Xing, R., Wong, S. H., & Liu, Y. (2016). Distinct subtypes of gastric cancer defined by molecular characterization include novel mutational signatures with prognostic capability, 76(7), 1724–1733. doi:10.1158/0008-5472.CAN-15-2443
}
\references{}
\keyword{datasets}
