\name{TARGET_db}
\alias{TARGET_db}
\title{TARGET_db}
\description{
	Broad institute TARGET - tumor alterations relevant for genomics-driven therapy.
	Downloaded from http://www.broadinstitute.org/cancer/cga/target.
}
\usage{data(TARGET_db)}
\format{
Two data frames:
1. TARGET_db_v2_05042014
2. TARGET_db_v1_07012013
}
\source{
}
\references{}
\keyword{datasets}
